Extract code into Field
